Details Icon
Some weird characters in alt tags were causing errors as no matching
icons could be found for some apps (eg. Weather+). I have changed it to
look only for icons that are 175x175 - I am well aware this is kludgy
and I'll be looking for a better solution later on.